+++ packages/cycle-scripts-one-fits-all/template/tsconfig.json
@@ -14,6 +14,7 @@
         "lib": ["ES6", "ES5", "DOM"]
     },
     "exclude": [
-        "node_modules"
+        "node_modules",
+        "build"
     ]
 }
